PRUEBAS/requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.2.0
exceptiongroup==1.2.0
fastapi==0.68.0
flake8==7.0.0
flake8-builtins==2.2.0
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.2
pluggy==1.3.0
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==7.4.4
tomli==2.0.1
requests==2.28.2
